订阅检测获取概念树展示

+++ kgms/src/main/java/com/plantdata/kgcloud/domain/dw/service/impl/GraphMapServiceImpl.java
@@ -15,12 +15,14 @@ import com.plantdata.kgcloud.domain.dw.rsp.GraphMapRsp;
 import com.plantdata.kgcloud.domain.dw.service.DWService;
 import com.plantdata.kgcloud.domain.dw.service.GraphMapService;
 import com.plantdata.kgcloud.domain.dw.service.PreBuilderService;
+import com.plantdata.kgcloud.domain.edit.service.ConceptService;
 import com.plantdata.kgcloud.domain.kettle.service.KettleLogStatisticService;
 import com.plantdata.kgcloud.exception.BizException;
 import com.plantdata.kgcloud.sdk.rsp.app.main.AttrExtraRsp;
 import com.plantdata.kgcloud.sdk.rsp.app.main.AttributeDefinitionRsp;
 import com.plantdata.kgcloud.sdk.rsp.app.main.BaseConceptRsp;
 import com.plantdata.kgcloud.sdk.rsp.app.main.SchemaRsp;
+import com.plantdata.kgcloud.sdk.rsp.edit.BasicInfoVO;
 import com.plantdata.kgcloud.util.ConvertUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Example;
@@ -70,6 +72,9 @@ public class GraphMapServiceImpl implements GraphMapService {
     private DWDatabaseRepository dwDatabaseRepository;
 
     @Autowired
+    private ConceptService conceptService;
+
+    @Autowired
     private DWService dwService;
     @Override
     public List<GraphMapRsp> list(String userId, GraphMapReq graphMapReq) {
@@ -87,12 +92,12 @@ public class GraphMapServiceImpl implements GraphMapService {
 
             if (graphMapList != null && !graphMapList.isEmpty()) {
 
-                List<DWGraphMap> childGraphMaps = new ArrayList<>();
-                for (DWGraphMap graphMap : graphMapList) {
-                    getChildConceptMap(childGraphMaps, graphMapReq.getKgName(), graphMapReq.getDatabaseId(), graphMap.getConceptId());
+                if(graphMapReq.getConceptId() != null){
+                    List<DWGraphMap> childGraphMaps = new ArrayList<>();
+                    getChildConceptMap(childGraphMaps, graphMapReq.getKgName(), graphMapReq.getDatabaseId(), graphMapReq.getConceptId());
+                    graphMapList.addAll(childGraphMaps);
                 }
 
-                graphMapList.addAll(childGraphMaps);
             }
 
         }
@@ -390,21 +395,30 @@ public class GraphMapServiceImpl implements GraphMapService {
 
     private void deleteConcept(DWGraphMap graphMap) {
 
-
         Integer attrId = graphMap.getAttrId();
 
         if(attrId != null){
             deleteRelationAttr(graphMap.getKgName(),attrId);
         }
 
-        graphMapRepository.deleteById(graphMap.getId());
+        synchronized (this){
+            if(graphMapRepository.existsById(graphMap.getId())){
+                graphMapRepository.deleteById(graphMap.getId());
+            }
+        }
     }
 
     private void deleteRelationAttr(String kgName, Integer attrId) {
 
         List<DWGraphMapRelationAttr> relationAttrList = graphMapRelationAttrRepository.findAll(Example.of(DWGraphMapRelationAttr.builder().kgName(kgName).attrId(attrId).build()));
         if(relationAttrList != null && !relationAttrList.isEmpty()){
-            relationAttrList.forEach(attr -> graphMapRelationAttrRepository.deleteById(attr.getId()));
+            relationAttrList.forEach(attr -> {
+                synchronized (this) {
+                    if (graphMapRelationAttrRepository.existsById(attr.getId())) {
+                        graphMapRelationAttrRepository.deleteById(attr.getId());
+                    }
+                }
+            });
         }
 
     }
@@ -489,17 +503,23 @@ public class GraphMapServiceImpl implements GraphMapService {
 
     private void getChildConceptMap(List<DWGraphMap> graphMapList, String kgName, Long databaseId, Long conceptId) {
 
-        List<DWGraphMap> graphMaps = graphMapRepository.findAll(Example.of(DWGraphMap.builder().kgName(kgName).dataBaseId(databaseId).pConceptId(conceptId).build()));
-        if (graphMaps == null || graphMaps.isEmpty()) {
+        List<BasicInfoVO> conceptTree = conceptService.getConceptTree(kgName, conceptId);
+
+        if(conceptTree ==  null || conceptTree.isEmpty()){
             return;
         }
 
-        for (DWGraphMap graphMap : graphMaps) {
-            graphMapList.add(graphMap);
-            if (graphMap.getConceptId() == null) {
+        for(BasicInfoVO info : conceptTree){
+            if(info.getId() == null || info.getId().equals(conceptId)){
                 continue;
             }
-            getChildConceptMap(graphMapList, kgName, databaseId, graphMap.getConceptId());
+
+            List<DWGraphMap> graphMaps = graphMapRepository.findAll(Example.of(DWGraphMap.builder().kgName(kgName).dataBaseId(databaseId).conceptId(info.getId()).build()));
+            if (graphMaps == null || graphMaps.isEmpty()) {
+                continue;
+            }
+
+            graphMapList.addAll(graphMaps);
         }
 
         return;
